fix conflicts and rollbacks
