+++ src/OpenCvSharp/Modules/photo/Tonemap.cs
@@ -1,5 +1,6 @@
 ﻿using System;
-// ReSharper disable UnusedMember.Global
+using System.Collections.Generic;
+using OpenCvSharp.Util;
 
 namespace OpenCvSharp
 {
@@ -8,42 +9,30 @@ namespace OpenCvSharp
     /// </summary>
     public class Tonemap : Algorithm
     {
-        private Ptr? ptrObj;
+        private Tonemap.Ptr? ptrObj;
 
         /// <summary>
-        /// Constructor used by Tonemap.Create
+        /// Creates instance by raw pointer cv::ml::Boost*
         /// </summary>
-        private Tonemap()
+        protected Tonemap(IntPtr p)
+            : base()
         {
-        }
-
-        /// <summary>
-        /// Constructor used by subclasses
-        /// </summary>
-        protected Tonemap(IntPtr ptr)
-        {
-            this.ptrObj = null;
-            this.ptr = ptr;
+            ptrObj = new Tonemap.Ptr(p);
+            ptr = ptrObj.Get();
         }
 
         /// <summary>
         /// Creates simple linear mapper with gamma correction
         /// </summary>
-        /// <param name="gamma">positive value for gamma correction. Gamma value of 1.0 implies no correction, gamma
-        /// equal to 2.2f is suitable for most displays.
+        /// <param name="gamma">gamma positive value for gamma correction.
+        /// Gamma value of 1.0 implies no correction, gamma equal to 2.2f is suitable for most displays.
         /// Generally gamma &gt; 1 brightens the image and gamma &lt; 1 darkens it.</param>
         /// <returns></returns>
-        public static Tonemap Create(float gamma = 1.0f)
+        public static Tonemap Create(float gamma = 2.2f)
         {
             NativeMethods.HandleException(
-                NativeMethods.photo_createTonemap(gamma, out var ptrObjPtr));
-
-            var ptrObj = new Ptr(ptrObjPtr);
-            return new Tonemap
-            {
-                ptrObj = ptrObj,
-                ptr = ptrObj.Get()
-            };
+                NativeMethods.photo_createTonemap(gamma, out var ptr));
+            return new Tonemap(ptr);
         }
 
         /// <summary>
@@ -59,23 +48,24 @@ namespace OpenCvSharp
         /// <summary>
         /// Tonemaps image
         /// </summary>
-        /// <param name="src">CV_32FC3 Mat (float 32 bits 3 channels)</param>
-        /// <param name="dst">CV_32FC3 Mat with values in [0, 1] range</param>
+        /// <param name="src">source image - CV_32FC3 Mat (float 32 bits 3 channels)</param>
+        /// <param name="dst">destination image - CV_32FC3 Mat with values in [0, 1] range</param>
         public virtual void Process(InputArray src, OutputArray dst)
         {
-            if (src == null) 
+            if (src == null)
                 throw new ArgumentNullException(nameof(src));
             if (dst == null)
                 throw new ArgumentNullException(nameof(dst));
+
             src.ThrowIfDisposed();
             dst.ThrowIfNotReady();
-
+            
             NativeMethods.HandleException(
                 NativeMethods.photo_Tonemap_process(ptr, src.CvPtr, dst.CvPtr));
 
+            GC.KeepAlive(this);
             GC.KeepAlive(src);
             dst.Fix();
-            GC.KeepAlive(this);
         }
 
         /// <summary>
@@ -102,7 +92,7 @@ namespace OpenCvSharp
             }
         }
 
-        private class Ptr : OpenCvSharp.Ptr
+        internal class Ptr : OpenCvSharp.Ptr
         {
             public Ptr(IntPtr ptr) : base(ptr)
             {
